websocket_usage_example.js
Note: at current state the cex.io websocket is not for public use intended, all data delievered there is actually for the website, but you can do atleast everything cex.io do (chat, charts, balance, etc.).

The example retrieve lowest ask, highest bid, and the orderbook (but this is not the complete version and is used for the website orderbook with the 50 order limit and the depth graph), at the console you will also see many other informations that is delieverd by the websocket e.g. chat messages.

The "ohlcv" things are for the chart, so if you want todo your own chart just put the cases down to the console.debug see what you get there.